fix(lens): the socket-holder fold moves to osfacts-client
`foldSocketHoldersReading`, `osfactsSocketHolders`, `SocketHolder` and
`SocketHolderReading` now live in `osfacts/client-ts/src/client.ts`, beside
`foldStartTimeReading` / `processIdentityAsync` — the same kind of operation
(spawn one verb, fold its wire-faithful reading into the honest domain answer
with an "absent" arm that must not collapse) at one home. The supervisor keeps
the SHAPE OF THE ASK: `ReadSocketHolders`, plus a re-export of the two answer
types so a consumer wiring `EndpointSpec` never reaches past the package.

- hickey H2 ≡ lowy F1 — the fold was osfacts-client's kind of work done one
  layer up: it read osfacts wire vocabulary (`facet === "socket_holders"`)
  inside a surface package, re-derived `ProcessRow` as a second
  `{pid, command}` type, and left the identical fold for the OTHER OS fact on
  the far side of a package boundary. A holder-wire change now edits two
  packages in one layer instead of three across two.

Layering verified, not assumed: `osfacts-client` stays zero-`@kolu` and
Node-builtins-only; `just check` (pnpm typecheck across every workspace +
biome) is green. As hickey noted, the `deps.closure.test.ts` allowlist entry
does NOT disappear — the guard counts type-only imports — so its comment is
re-aimed rather than deleted. The fold's six unit pins move with it into
`client.test.ts` (client-ts 28 → 34 tests, supervisor 136 → 130); the
supervisor's file keeps the gated real-binary pins, which are what it can
still prove. Public API surface: `ref-surface-supervisor.mdx` updated for the
two moved exports.

Raised by the lowy ∥ hickey lens review. Not pushed or merged.
