+++ website/src/content/surface/ref-surface-supervisor.mdx
@@ -56,8 +56,8 @@ It is generic over the client `C` and identity `I`.
 | `survivableSpawnDriver(cfg, deps?)` | the default `DaemonDriver`: `systemd-run --user` under a service, detached `+unref` otherwise |
 | `recycle(endpoint, steps)` · `serializeRestart(endpoint)` | `recycle` is the composed `capture → drain → recycle → reattach` sequence; `serializeRestart` returns a single-flight `triggerRestart` wrapper |
 | `waitForPidGone(pid, opts?)` | poll `isHolderLive` until the pid is reaped (`ESRCH`), so a respawn never races a live gate holder |
-| `osfactsSocketHolders(bin)` → `ReadSocketHolders` | the `readSocketHolders` inject, bound to an already-resolved osfacts binary path (`bakedOsFactsBin(<your env var>)` — this package never learns which var a consumer bakes). Calls `osfacts socket-holders <path> --procs`; the platform split, the parse, and the honesty policy live in that binary, not here. The identity source the gate-less-squatter recovery uses when the gate no longer names the holder. A leaf beside `waitForPidGone` — pid lifecycle via the OS |
-| `SocketHolderReading` · `foldSocketHoldersReading(reading)` | the **three-way** answer, never one possibly-empty list: `{ kind: "holders", holders }` (at least one named `{ pid, command }`), `{ kind: "none" }` (**proven** nothing holds it — only linux's world-readable `/proc/net/unix` can prove this), `{ kind: "unattributed", detail }` (something may hold it and the OS would not say what — a bound-but-unreadable holder, or a search that could not complete). Only `none` is evidence of freedom, and the endpoint still re-probes before spawning. `foldSocketHoldersReading` is that fold over an `osfacts-client` `SocketHoldersReading`, exported for consumers that spawn the binary themselves |
+| `ReadSocketHolders` | the TYPE of the `readSocketHolders` inject: `(socketPath: string) => Promise<SocketHolderReading>`. The reader that satisfies it is `osfactsSocketHolders(bin)` from **`osfacts-client`**, bound to an already-resolved osfacts binary path (`bakedOsFactsBin(<your env var>)` — this package never learns which var a consumer bakes). It calls `osfacts socket-holders <path> --procs`; the platform split, the parse, the honesty policy AND the fold live in the tool and its client, not here. Bind it to the same resolved path as `processIdentityAsync` so one root spells its env var once for both OS facts |
+| `SocketHolderReading` · `SocketHolder` | the **three-way** answer, never one possibly-empty list: `{ kind: "holders", holders }` (at least one named `{ pid, command }`), `{ kind: "none" }` (**proven** nothing holds it — only linux's world-readable `/proc/net/unix` can prove this), `{ kind: "unattributed", detail }` (something may hold it and the OS would not say what — a bound-but-unreadable holder, or a search that could not complete). Only `none` is evidence of freedom, and the endpoint still re-probes before spawning. Both types are **re-exported from `osfacts-client`**, whose `foldSocketHoldersReading(reading)` produces them from a `SocketHoldersReading` — one fold, so kolu and drishti cannot spell the three-way differently |
 | `SocketSquatterForeignError` · `isSocketSquatterForeignError(err)` | thrown when the process holding the rendezvous socket did NOT complete the kaval handshake — a foreign / non-conforming holder that is **never** killed. Carries `socketPath` and `holders: SocketHolder[]` (each pid + command) so the refusal names the squatter. Brand-checked, realm-robust |
 | `SocketProbeIndeterminateError` · `isSocketProbeIndeterminateError(err)` | thrown when a one-shot socket connect probe times out (`SocketServeState` `"indeterminate"`) — not proof free or dead. The endpoint never kills, unlinks, or spawns on it; carries `gatePath` / `socketPath` / optional `gatePid`. Failures after `connecting` emit `dead` then rethrow so status never wedges |
 | `DaemonContractSkewError` · `isContractSkewError(err)` | the ONE typed connect rejection that proves incompatibility. Constructed from fields — `{ subject, daemonVersion, requiredVersion, pid? }` — with the message derived from them (no consumer ever re-parses the prose); `readonly daemonVersion` / `requiredVersion` ride the instance so downstream translators (a typed RPC rethrow, a status arm) read them structurally. The optional `readonly pid` is the skewed daemon's self-reported OS pid (additive), carried so the gate-less-squatter recovery of an OLD orphan — which throws here before a connection exists — has its identity attestation. The guard is brand-checked (realm-robust), never `instanceof` |
